server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.20.0
pillow==11.0.0
scipy==1.11.4
pymongo==4.6.1
python-dotenv==1.0.0

